refactor: split agents page components
